requirements.txt
click
vivo-rdflib-sparqlstore==0.2
# only needed for sample job
requests